Initial support for request bodies
This is very limited support. It introduces using "triple tick" syntax to delimit the body content.

```
body...
```

This simplifies lexing and hopefully is temporary. Might be a limitation
of abusing lalrpop and logos

Bodies can also only be a single line. Logos detects newlines \n as
newline tokens even when a higher priority is set on the body token
regex